Use datatype of text so substitutions and html works
The PCR Console was designed with plain text responses from the PCR API in mind.
However, when we upgraded the PCR API to set the content type correctly to JSON,
jQuery automatically started transforming the data into a JSON object, wreaking
havoc on the regex and HTML insertion in the console. By explicitly setting the
dataType of the ajax query to "text", we can get the original behavior again.

Fixes #1